% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{acres_to_kmsq}
\alias{acres_to_kmsq}
\title{Convert acres to km squared.}
\usage{
acres_to_kmsq(acres)
}
\arguments{
\item{acres}{numeric acres to be converted.}
}
\value{
returns kilometers squared.
}
\description{
This is a whole paragraph explaining what this function does. For
reference see \code{https://google.com/acres_to_kmsq}.
}
